fetch/metadata/generated/element-input-image.https.sub.html

<!DOCTYPE html>
<!--
This test was procedurally generated. Please do not modify it directly.
Sources:
- fetch/metadata/tools/fetch-metadata.conf.yml
- fetch/metadata/tools/templates/element-input-image.sub.html
-->
<html lang="en">
  <meta charset="utf-8">
  <title></title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/fetch/metadata/resources/helper.sub.js"></script>
  <body>
  <script>
  'use strict';

  function induceRequest(url, test) {
    const input = document.createElement('input');
    input.setAttribute('type', 'image');

    document.body.appendChild(input);
    test.add_cleanup(() => input.remove());

    return new Promise((resolve) => {
        input.onload = input.onerror = resolve;
        input.setAttribute('src', url);
      });
  }

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(makeRequestURL(key, ['sameOrigin', 'crossSite', 'sameOrigin']), t)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'cross-site');
        });
  }, 'sec-fetch-site - Same-Origin -> Cross-Site -> Same-Origin redirect, no attributes');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(makeRequestURL(key, ['crossSite', 'sameOrigin']), t)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'cross-site');
        });
  }, 'sec-fetch-site - Cross-Site -> Same Origin, no attributes');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(makeRequestURL(key, ['crossSite', 'sameSite']), t)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'cross-site');
        });
  }, 'sec-fetch-site - Cross-Site -> Same-Site, no attributes');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(makeRequestURL(key, ['crossSite', 'crossSite']), t)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'cross-site');
        });
  }, 'sec-fetch-site - Cross-Site -> Cross-Site, no attributes');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(makeRequestURL(key, ['sameOrigin', 'sameOrigin']), t)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'same-origin');
        });
  }, 'sec-fetch-site - Same-Origin -> Same Origin, no attributes');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(makeRequestURL(key, ['sameOrigin', 'sameSite']), t)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'same-site');
        });
  }, 'sec-fetch-site - Same-Origin -> Same-Site, no attributes');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(makeRequestURL(key, ['sameOrigin', 'crossSite']), t)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'cross-site');
        });
  }, 'sec-fetch-site - Same-Origin -> Cross-Site, no attributes');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(makeRequestURL(key, ['sameSite', 'sameOrigin']), t)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'same-site');
        });
  }, 'sec-fetch-site - Same-Site -> Same Origin, no attributes');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(makeRequestURL(key, ['sameSite', 'sameSite']), t)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'same-site');
        });
  }, 'sec-fetch-site - Same-Site -> Same-Site, no attributes');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(makeRequestURL(key, ['sameSite', 'crossSite']), t)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'cross-site');
        });
  }, 'sec-fetch-site - Same-Site -> Cross-Site, no attributes');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(makeRequestURL(key, ['sameOrigin', 'httpOrigin', 'sameOrigin']), t)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'cross-site');
        });
  }, 'sec-fetch-site - HTTPS downgrade-upgrade, no attributes');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(makeRequestURL(key, []), t)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-mode');
          assert_equals(headers['sec-fetch-mode'], 'no-cors');
        });
  }, 'sec-fetch-mode - no attributes');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(makeRequestURL(key, []), t)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-dest');
          assert_equals(headers['sec-fetch-dest'], 'image');
        });
  }, 'sec-fetch-dest - no attributes');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(makeRequestURL(key, []), t)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_not_own_property(headers, 'sec-fetch-user');
        });
  }, 'sec-fetch-user - no attributes');
  </script>
  </body>
</html>
